pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "dnspython>=2.8.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
